init_app.rsからサブモジュールを切り出し
